-) Correzione del calclolo della pressione media -) MAX_PLATE_TEMP non più define ma variabile calcolata a giro di programma con la updateMaxTempPlate() in base all'algoritmo di SB -) MIN_PLATE_TEMP non più define ma variabile calcolata a giro di programma con la updateMinTempPlate() in base all'algoritmo di SB -) valori di temperatura mandati alla PRO gestiti con segno -) migliorato algoritmo di ritrasmissione messaggio in caso d mancato stp delle pompe -) gestito comne allarme la ricezione del valore '255' dai sensori di flusso sul cmapo ARIA

+++ SCTRO2_C/Application/App_Ges_Func.c
@@ -1294,7 +1294,7 @@ void PlateStateFrigo(void)
 			break;
 		case PLATE_STATE_DIS_FRIGO:
 			// controllo disabilitazione raffreddamento
-			if(T_PLATE_C_GRADI_CENT < (float)MIN_PLATE_TEMP)
+			if(T_PLATE_C_GRADI_CENT < MIN_PLATE_TEMP)
 			{
 				if(timeIntervalPlate10 && (msTick10_elapsed(timeIntervalPlate10) >= 300))
 				{
@@ -1311,7 +1311,7 @@ void PlateStateFrigo(void)
 			break;
 		case PLATE_STATE_EN_FRIGO:
 			// controllo abilitazione raffreddamento
-			if(T_PLATE_C_GRADI_CENT > (float)MIN_PLATE_TEMP)
+			if(T_PLATE_C_GRADI_CENT > MIN_PLATE_TEMP)
 			{
 				if(timeIntervalPlate10 && (msTick10_elapsed(timeIntervalPlate10) >= 300))
 				{
@@ -1340,7 +1340,7 @@ void PlateStateHeating(void)
 			break;
 		case PLATE_HEAT_DIS:
 			// controllo disabilitazione riscaldamento, raffreddamento
-			if(T_PLATE_C_GRADI_CENT > (float)MAX_PLATE_TEMP)
+			if(T_PLATE_C_GRADI_CENT > MAX_PLATE_TEMP)
 			{
 				if(timeIntervalPlate10_r && (msTick10_elapsed(timeIntervalPlate10_r) >= 300))
 				{
@@ -1355,7 +1355,7 @@ void PlateStateHeating(void)
 			break;
 		case PLATE_HEAT_EN:
 			// controllo abilitazione riscaldamento, raffreddamento
-			if(T_PLATE_C_GRADI_CENT < (float)MAX_PLATE_TEMP)
+			if(T_PLATE_C_GRADI_CENT < MAX_PLATE_TEMP)
 			{
 				if(timeIntervalPlate10_r && (msTick10_elapsed(timeIntervalPlate10_r) >= 300))
 				{
